This commit (I think) corrects the calculation of the density field for TALA cases (I need to verify that I'm correct with Cian, but wanted to run through buildbot - I can revert if I'm wrong).
In essence, previously, density was calculated incorreclty, as values were being scaled with pressure (for TALA simulations, the pressure effect on density, via the bulk modulus, is ignored). For TALA simulations where we use the reference density in the continuity equation (which is what we usually do) this has previously not made any difference, as it's the buoyancy density which drives the flow, and this was always calculated correctly. However, if we use the full density in the continuity equation, then it is an issue.

I hope this commit fixes that, but would be keen to discuss with Cian tomorrow (pelase don't shout!).