Tidying up latent heating.
A pointwise algorithm is now available so that no projection is necessary if the mesh is suitable.

Default the clapeyron slope and delta rho to those found elsewhere if they are unspecified beneath the algorithm.

Fix a sign error in the absorption that was the source of all my benchmarking woes.....