Add an optional reference_temperature at which the reference_depth will be interpretted.
This allows the reference depth to be set to a constant and have it not move if the clapeyron slope changes so long as the reference temperature is known.

Adjust the tests to test for this.

Also adjust the latent heating so it is consistent with being used in a MantleAnelasticEnergy equation type.