Always create %region_ids on extruded mesh,
even if the horizontal mesh doesn't have a %region_ids. This fixes the
case where extruded_region_ids are specified in one or more layers, but
the horizontal input mesh does not have any. To keep things simple the
extruded mesh now always allocates a region_ids array which if
extruded_region_ids are not specified, and there are no horizontal
region_ids to copy from, default to 0.