Introduce monotonic P2 shape
This is an idea of @jrper to use a local change of basis to obtain basis
functions that are strictly positive, monotonic and take their maximum value in
their associated node. This is done by replacing the Lagrangian P2 basis functions
associated with the vertex nodes with a linear combination of it with
the basis functions associated with the adjacent edge nodes. The
resulting basis is no longer nodal because the new "vertex" basis
functions are nonzero in the adjacent edges as well. Thus we no longer
have an equivalence of coefficients of basis functions with their values
in associated nodes. By construction, the change of each vertex basis
function only affects the elements adjacent to that vertex so we can
still construct piecewise continuous P2 functions in exactly the same
way as with the standard Lagrangian P2 basis.

This might be useful for P2 mass lumping, and strict bounding of P2
fields.

The distinction between a standard Lagrangian P2 shape and the
monotonic P2 shape is made by setting a new logical attribute %nodal,
which defaults to .true. and is currently only set to .false. for the
new monotonic P2 shape. Note that we are not using the element_family
entry on the shape numbering, which will still be set to
ELEMENT_LAGRANGIAN for both P2 shapes as they both use the same node
numbering. The %nodal flag is picked up in remap_field (and
remap_field_to_surface) which if necessary not just maps from
"from_field" coefficients to "to_field" nodal values, but also maps
these nodal values to "to_field" coefficients. Thus using fields based
on the new monotonic P2 shape can be used as from_field or to_field in
all cases where this is currently possible with standard Lagrangian P2
fields.

For the moment, this is only coded for use in the code (i.e. no schema,
populate_state, make_mesh(), make_element_shape() stuff) where a
monotonic P2 mesh can be created on the fly from an existing
Lagrangian P2 mesh by calling lagrangian_to_monotonic_p2_mesh(), after
which existing P2 fields can be mapped between the two meshes with
remap_field().