Clean up of "SOME TEMPORARY DEBUGGING STUFF" commit
Keeping the debugging output for inverted elements, but now abort the
run if any are found. This commit also contained a fix that ensures the
basegeometry is redistributed (and not dropped) by zoltan. The extra
adapted_quality output seems unnecessary as it should be the same as the
one directly after coming out of libadaptivity (before interpolation and
redistribution.)